refactor build and publish subcommand to use common base class
