+++ src/compose_flow/commands/subcommands/publish.py
@@ -1,40 +1,11 @@
-import logging
 
-from functools import lru_cache
-
-from .base import BaseSubcommand
+from .base import BaseBuildSubcommand
 
 
-class Publish(BaseSubcommand):
+class Publish(BaseBuildSubcommand):
     """
     Subcommand for building and pushing Docker images
     """
-
-    rw_env = True
-    remote_action = True
-    update_version_env_vars = True
-
-    def build(self):
-        compose = self.compose
-
-        compose.handle(extra_args=['build'])
-
-    @property
-    @lru_cache()
-    def compose(self):
-        """
-        Returns a Compose subcommand
-        """
-        from .compose import Compose
-
-        return Compose(self.workflow)
-
-    def do_validate_profile(self):
-        return False
-
-    @classmethod
-    def fill_subparser(cls, parser, subparser) -> None:
-        pass
 
     def get_built_docker_images(self) -> list:
         """
@@ -49,18 +20,6 @@ class Publish(BaseSubcommand):
 
         return list(docker_images)
 
-    def handle(self):
-        self.build()
-
-        self.push()
-
-    def is_missing_env_arg_okay(self):
-        return True
-
-    @property
-    def logger(self):
-        return logging.getLogger(f'{__name__}.{self.__class__.__name__}')
-
     def push(self):
         docker_images = self.get_built_docker_images()
         for docker_image in docker_images:
@@ -71,3 +30,8 @@ class Publish(BaseSubcommand):
                 self.logger.info(f'docker push {docker_image}')
             else:
                 self.execute(f'docker push {docker_image}', _fg=True)
+
+    def handle(self):
+        self.build(pull=False)
+
+        self.push()
